backend/requirements.txt
langchain
langchain-openai
python-dotenv
fastapi
python-multipart
beautifulsoup4
requests
selenium
pandas
PyMuPDF
chromadb
tiktoken
langchain-community
langchain-text-splitters